Core food web model

Core food web model

FoodWeb_SQO(NumSim, csed, cwater, cpw, log_KowTS, logkow_tempcor, EdA, EdB,
  xdoc, ddoc, xpoc, dpoc, alphapoc, alphadoc, ocsed, ds, taxa, A, B, T,
  lipid, nloc, nlom, wc, beta, betap, mo, mp, phi, kM, Wb, Cox, vss, scav,
  preyprop, cbiota, vld, vcd, vnd, vwd, GR, assimEff_1, assimEff_2,
  assimEff_3)

Arguments

NumSim
required input

csed
required input

cwater
required input

cpw
required input

log_KowTS
required input

logkow_tempcor
required input

EdA
required input

EdB
required input

xdoc
required input

ddoc
required input

xpoc
required input

dpoc
required input

alphapoc
required input

alphadoc
required input

ocsed
required input

ds
required input

taxa
required input

A
required input

B
required input

T
required input

lipid
required input

nloc
required input

nlom
required input

wc
required input

beta
required input

betap
required input

mo
required input

mp
required input

phi
required input

kM
required input

Wb
required input

Cox
required input

vss
required input

scav
required input

preyprop
required input

cbiota
required input

vld
required input

vcd
required input

vnd
required input

vwd
required input

GR
required input

assimEff_1
required input

assimEff_2
required input

assimEff_3
required input